move schedule sync into zuluru sync class
