+++ server/lib/__init__.py
@@ -4,4 +4,3 @@ from .teams_serializer import build_teams_response
 from .players_serializer import build_players_response
 from .zuluru_sync import ZuluruSync
 from .player_db import PlayerDb
-from .schedule_sync import ScheduleSync
